+++ pacman-app-arcade-pacman/src/main/java/de/amr/games/pacman/arcade/pacman/ArcadePacMan_GameRenderer.java
@@ -11,8 +11,8 @@ import de.amr.games.pacman.model.actors.Bonus;
 import de.amr.games.pacman.ui2d.GameRenderer;
 import de.amr.games.pacman.ui2d.assets.AssetStorage;
 import de.amr.games.pacman.ui2d.assets.GameSpriteSheet;
-import javafx.beans.property.DoubleProperty;
-import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.FloatProperty;
+import javafx.beans.property.SimpleFloatProperty;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.paint.Color;
 
@@ -29,7 +29,7 @@ public class ArcadePacMan_GameRenderer implements GameRenderer {
 
     private final AssetStorage assets;
     private final ArcadePacMan_SpriteSheet spriteSheet;
-    private final DoubleProperty scalingPy = new SimpleDoubleProperty(1.0);
+    private final FloatProperty scalingPy = new SimpleFloatProperty(1.0f);
     private final Canvas canvas;
     private boolean mazeHighlighted;
     private boolean blinkingOn;
@@ -67,7 +67,7 @@ public class ArcadePacMan_GameRenderer implements GameRenderer {
     }
 
     @Override
-    public DoubleProperty scalingProperty() {
+    public FloatProperty scalingProperty() {
         return scalingPy;
     }
